Add updated labels component to Experiment Detail
