--- client/app/containers/Label/constants.js
@@ -1,20 +1,0 @@
-export const LOAD_EXPERIMENT_LABELS =
-  'app/ExperimentDetail/LOAD_EXPERIMENT_LABELS';
-export const LOAD_EXPERIMENT_LABELS_SUCCESS =
-  'app/ExperimentDetail/LOAD_EXPERIMENT_LABELS_SUCCESS';
-export const LOAD_EXPERIMENT_LABELS_ERROR =
-  'app/ExperimentDetail/LOAD_EXPERIMENT_LABELS_ERROR';
-
-export const CREATE_EXPERIMENT_LABELS =
-  'app/ExperimentDetail/CREATE_EXPERIMENT_LABELS';
-export const CREATE_EXPERIMENT_LABELS_SUCCESS =
-  'app/ExperimentDetail/CREATE_EXPERIMENT_LABELS_SUCCESS';
-export const CREATE_EXPERIMENT_LABELS_ERROR =
-  'app/ExperimentDetail/CREATE_EXPERIMENT_LABELS_ERROR';
-
-export const DELETE_EXPERIMENT_LABELS =
-  'app/ExperimentDetail/DELETE_EXPERIMENT_LABELS';
-export const DELETE_EXPERIMENT_LABELS_SUCCESS =
-  'app/ExperimentDetail/DELETE_EXPERIMENT_LABELS_SUCCESS';
-export const DELETE_EXPERIMENT_LABELS_ERROR =
-  'app/ExperimentDetail/DELETE_EXPERIMENT_LABELS_ERROR';
